Set correct starter on imported sites and moved more OCW code
