+++ README.md
@@ -3,8 +3,9 @@ OCW Studio manages deployments for OCW courses.
 
 **SECTIONS**
 1. [Initial Setup](#initial-setup)
-1. [Optional Setup](#optional-setup)
-1. [Running Tests](#running-tests)
+1. [Testing and Formatting](#testing-and-formatting)
+1. [Importing OCW course sites](#importing-ocw-course-sites)
+
 
 # Initial Setup
 
@@ -12,59 +13,25 @@ ocw_studio follows the same [initial setup steps outlined in the common ODL web 
 Run through those steps **including the addition of `/etc/hosts` aliases and the optional step for running the
 `createsuperuser` command**.
 
-# Optional Setup
+### Commits
 
-Described below are some setup steps that are not strictly necessary
-for running the app
-
-# Running tests
-
-*NOTE: These commands can be run with ```docker-compose exec``` to execute them
-in an already-running container, or with ```docker-compose run --rm``` to
-execute them in a new container.*
-
-## Python Tests/Linting
-
-
-To run Python tests:
-
-```sh
-docker-compose run --rm web pytest
+To ensure commits to github are safe, you should install the following first:
+```
+pip install pre_commit
+pre-commit install
 ```
 
-To run Python tests in a single file:
-
-```sh
-docker-compose run --rm web pytest /path/to/test.py
+To automatically install precommit hooks when cloning a repo, you can run this:
+```
+git config --global init.templateDir ~/.git-template
+pre-commit init-templatedir ~/.git-template
 ```
 
-To run Python test cases in a single file that match some function/class name:
 
-```sh
-docker-compose run --rm web pytest /path/to/test.py -k test_some_logic
-```
+# Testing and Formatting
 
-To run the Python linter:
-
-```sh
-docker-compose run --rm web pylint
-```
-
-## Python Formatting
-
-We use [black](https://github.com/psf/black) to format our Python code.
-
-To format all python files:
-
-```sh
-docker-compose run --rm web black .
-```
-
-To format a specific file:
-
-```sh
-docker-compose run --rm web black /path/to/file.py
-```
+Writing tests, running the test suite, and formatting code follows the same steps that are outlined in [the common ODL web app guide](https://github.com/mitodl/handbook/blob/master/common-web-app-guide.md#testing-and-formatting).
+Below are some steps that may be particular to this project.
 
 ## JS/CSS Tests and Linting
 
@@ -136,57 +103,25 @@ npm run fmt
 
 You can also try `npm run fmt:check` to see if any files need to be reformatted.
 
-# Running the app in a notebook
 
-This repo includes a config for running a [Jupyter notebook](https://jupyter.org/) in a
-Docker container. This enables you to do in a Jupyter notebook anything you might
-otherwise do in a Django shell. To get started:
+# Importing OCW course sites
 
-- Copy the example file
-    ```bash
-    # Choose any name for the resulting .ipynb file
-    cp localdev/app.ipynb.example localdev/app.ipynb
-    ```
-- Build the `notebook` container _(for first time use, or when requirements change)_
-    ```bash
-    docker-compose -f docker-compose-notebook.yml build
-    ```
-- Run all the standard containers (`docker-compose up`)
-- In another terminal window, run the `notebook` container
-    ```bash
-    docker-compose -f docker-compose-notebook.yml run --rm --service-ports notebook
-    ```
-- Visit the running notebook server in your browser. The `notebook` container log output will
-  indicate the URL and `token` param with some output that looks like this:
-    ```
-    notebook_1  |     To access the notebook, open this file in a browser:
-    notebook_1  |         file:///home/mitodl/.local/share/jupyter/runtime/nbserver-8-open.html
-    notebook_1  |     Or copy and paste one of these URLs:
-    notebook_1  |         http://(2c19429d04d0 or 127.0.0.1):8080/?token=2566e5cbcd723e47bdb1b058398d6bb9fbf7a31397e752ea
-    ```
-  Here is a one-line command that will produce a browser-ready URL from that output. Run this in a separate terminal:
-    ```bash
-    docker logs $(docker ps --format '{{.Names}}' | grep "_notebook_run_") | grep -E "http://(.*):8080[^ ]+\w" | tail -1 | sed -e 's/^[[:space:] or]*//'
-    ```
-  OSX users can pipe that output to `xargs open` to open a browser window directly with the URL from that command.
-- Navigate to the `.ipynb` file that you created and click it to run the notebook
-- Execute the first block to confirm it's working properly (click inside the block
-  and press Shift+Enter)
+We have raw data for numerous course sites in cloud storage. These sites can be imported into OCW Studio and 
+saved into the database via a management command: `import_ocw_course_sites`.
 
-From there, you should be able to run code snippets with a live Django app just like you
-would in a Django shell.
-
-
-# Commits
-
-To ensure commits to github are safe, you should install the following first:
+This command will only work if you have the following:
+1. The name of the AWS bucket which contains the course data. Ask a fellow developer for this.
+1. Valid settings for `AWS_ACCESS_KEY_ID` and `AWS_SECRET_ACCESS_KEY`. As with other settings, these should be
+   specified in your `.env` file. You can copy these from heroku, or ask a fellow developer for them.
+   
+Some example commands:
+```bash
+# List the data file names of all available course sites
+manage.py import_ocw_course_sites -b <bucket_name> --list
+# List the data file names of all course sites that match a filter
+manage.py import_ocw_course_sites -b <bucket_name> --filter frameworks-of-urban-governance --list 
+# Import course sites with a data file name that matches a filter
+manage.py import_ocw_course_sites -b <bucket_name> --filter frameworks-of-urban-governance 
+# Import ALL course sites (this will take quite a while)
+manage.py import_ocw_course_sites -b <bucket_name>
 ```
-pip install pre_commit
-pre-commit install
-```
-
-To automatically install precommit hooks when cloning a repo, you can run this:
-```
-git config --global init.templateDir ~/.git-template
-pre-commit init-templatedir ~/.git-template
-```
