+++ ocw_import/management/commands/import_ocw_course_sites.py
@@ -1,4 +1,4 @@
-""" Backpopulate OCW courses and content via ocw2hugo output """
+""" Import OCW course sites and content via ocw2hugo output """
 import pydoc
 
 from django.core.management import BaseCommand
@@ -9,9 +9,9 @@ from ocw_import.tasks import import_ocw2hugo_courses
 
 
 class Command(BaseCommand):
-    """ Backpopulate OCW courses and content via ocw2hugo output """
+    """ Import OCW course sites and content via ocw2hugo output """
 
-    help = "Backpopulate OCW courses and content via ocw2hugo output"
+    help = __doc__
 
     def add_arguments(self, parser):
         parser.add_argument(
